add frontend and NNID fetching
this is the checkin of the basic frontend I've been working on for a few
weeks and now it's incorporated into the Go server, enjoy.
there are also NNID fetching routines thrown on the back of this, I
think next commit or the one after that will be a brief re-structure
just to move everything out of processor-server.go.
the NNID fetches are cached though it requests in JSON which, for
instance Pretendo account-server isn't compatible with as of this commit